+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased
 - [**BC**] Consume with error handling
 - Fix to work without a tracing
+- [**BC**] Remove `Logger` type and use `ILogger` from `Microsoft.Extensions.Logging`
 
 ## 16.2.0 - 2021-09-14
 - Update dependencies